Initial commit for Postfix puppet module
It took me a while to get to this point, I originally went
symlink central as described in http://rspec-puppet.com/setup/
but then after reading some more, and particularly in the need to
load in external dependencies that a .fixtures.yml file was the way
to go.  Anyway here it is, zero functionality but two tests that
pass!